MSLI-4 - change property to Int32 to match CoreData property so we can pass the series ID to search details

+++ My Series List/My Series List/Model/SeriesModel.swift
@@ -14,7 +14,7 @@ enum Api {
     }
     
     class Series: Codable {
-        let id: Double?
+        let id: Int32?
         let name: String?
         let overView: String?
         let posterPath: String?
